Add getJobsPostedByUser endpoint in jobController.js
This commit adds a new endpoint, getJobsPostedByUser, in the jobController.js file. This endpoint allows users to fetch the jobs they have posted. It retrieves the jobs from the database based on the jobProviderId associated with the user. If successful, the jobs are returned in the response. If there is an error, a 500 status code is returned along with an error message.